Semester-4/DigitalAnalytics/Lab3/requirements.txt
torch
torchvision
pillow
scikit-learn
tqdm
numpy
